# docs/powerusyservice/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a backend web service created for Powerusy operations.",
        "title": "Powerusy backend service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alao Ramon, Nabil Abubakar",
            "url": "http://www.swagger.io/support",
            "email": "alao.adebusy@gmail.com, nabbo247@yahoo.com"
        },
        "license": {
            "name": "AddyTech Solution Ltd",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8060",
    "basePath": "/",
    "paths": {
...
        "/api/company/ApproveKYC": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "Approve KYC document.",
                "parameters": [
                    {
                        "description": "KYC approval",
                        "name": "KYC",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.KYCApprovalIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMessage"
                        }
                    }
                }
            }
        },
...
        "/api/company/RegisterCompany": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "Register new company",
                "parameters": [
                    {
                        "description": "Company details",
                        "name": "company",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CompanyDetailsIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/company/UploadCompanyDocuments": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "Upload company document",
                "parameters": [
                    {
                        "description": "Upload company document",
                        "name": "CompanyDoc",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ImportationDocumentIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/company/UploadKYC": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "Upload KYC document.",
                "parameters": [
                    {
                        "description": "KYC document",
                        "name": "KYC",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.KYCIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CompanyDetailsIn": {
            "type": "object",
            "properties": {
                "AccountNumber": {
                    "type": "string"
                },
                "Approvalcomment": {
                    "type": "string"
                },
                "BankName": {
                    "type": "string"
                },
                "CompanyAddress": {
                    "type": "string"
                },
                "CompanyLocation": {
                    "type": "string"
                },
                "CompanyLogo": {
                    "type": "string"
                },
                "CompanyName": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "PostAddress": {
                    "type": "string"
                },
                "ServiceId": {
                    "type": "integer"
                },
                "StatusId": {
                    "type": "integer"
                },
                "UserId": {
                    "type": "integer"
                },
                "WorkingDays": {
                    "type": "string"
                },
                "WorkingHours": {
                    "type": "string"
                }
            }
        },
...
        "models.ImportationDocumentIn": {
            "type": "object",
            "properties": {
                "DocumentName": {
                    "type": "string"
                },
                "DocumentPath": {
                    "type": "string"
                },
                "ImportationId": {
                    "type": "string"
                }
            }
        },
        "models.KYCApprovalIn": {
            "type": "object",
            "properties": {
                "Approvalcomment": {
                    "type": "string"
                },
                "Approvedby": {
                    "type": "integer"
                },
                "Registeredid": {
                    "type": "integer"
                },
                "Status": {
                    "type": "integer"
                }
            }
        },
        "models.KYCIn": {
            "type": "object",
            "properties": {
                "Articlesofassociation": {
                    "type": "string"
                },
                "Auditedfinancialstatement": {
                    "type": "string"
                },
                "Certificateofincorporation": {
                    "type": "string"
                },
                "Memorandomofassociation": {
                    "type": "string"
                },
                "Powerofattorneygranted": {
                    "type": "string"
                },
                "Registeredid": {
                    "type": "integer"
                },
                "Taxclearancecertificate": {
                    "type": "string"
                },
                "Validbusinesslicense": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseMessage": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
...
    }
}